Fix bug in caching engine: terminate() should return to uninit state.
NNCache::terminate() is documented as returning the cache to the
uninitialized state, in which getBlob() and setBlob() do not touch
the cache.  However, terminate() was not setting mInitialized back
to false, and therefore getBlob() and setBlob() could access the
cache normally.

This bug originated in the opengl version of the engine that the nn
version of the engine was copied from.

Bug: 63905942

Test: aosp_angler-eng
      nn/cache/nnCache mm, run nnCache_test on angler

Change-Id: Id80f704ba8017e36ad774be6ae194ee9769abce0