nnCache API change: Stop hardcoding cache size parameters (key, value, total).
Now those values must be passed to nnCache::initialize().

Incidental change: Initialize previously-uninitialized field
nnCache::mSavePending in constructor.  This makes no functional
difference -- the desired value is false, and the only nnCache
instance is static, so the field would have been initialized to
false anyway.  Nonetheless, explicit initialization is good
practice, and that's what we already do for the field
mInitialized, which is likewise explicitly initialized to false.

Bug: 63905942

Test: aosp_angler-eng
      nn/cache/nnCache mm, run nnCache_test on angler

Change-Id: I00b557686055dd5f142ce5d2033ddf0d77e5fd89