FullyConnected: fix a memory leak
Clang's static analyzer was complaining that we'd leak this value if
activation wasn't one of kActivation{None,Relu,Relu6}. In reality, it
looks like we always leak it. To avoid leakage, use a stack allocation
instead.

An alternate approach would be to make it a function-local static, so
that threads aren't created+destroyed every time we call this function.
convQuant8 does this.

Since I'm not actually familiar with this code and the multithreading
guarantees it makes (if any), I took the easy way out.

Bug: 27101951
Test: mma. Memory leak complaint no longer appears.
Change-Id: I6193287a9bb14566faa9e4480413d428e8f717c1