Use ExecutionPlan to control (possibly multiple step) execution.
- Split StepExecutor out from ExecutionBuilder.  ExecutionBuilder
  represents the description of an execution passed in through the
  NNAPI; StepExecutor is used to perform a single "step" of that
  execution on a single device (or CPU).

- Introduce ExecutionPlan::Controller class.  This is a token used
  to allow ExecutionBuilder::startCompute() to iterate over a sequence
  of steps provided by ExecutionPlan, obtaining a StepExecutor for each
  one.

There should be no change from previous behavior; this is code
reorganization in preparation for supporting execution of partitioned
models.  While ExecutionPlan is still only used when the property
debug.nn.partition.test is set to 0, and it still controls the
execution only when that property is set to 2, we do use the new
StepExecutor always.

Bug: 63905942
Test: mma (user)
      mma (userdebug)
      ml/nn/runtime/tests (userdebug)
      (with debug.nn.partition.test 0, 1, and 2;
       no new failures, logcat looks plausible)

Change-Id: I17473210bcc5de40603d439ec49374030d1b8f36