Add ExecutionPlan::Controller support for Compound plans.
This includes proper handling of ExecutionBuilder inputs
and outputs, but not of inter-partition temporaries that
need to become inputs and outputs.

When debug.nn.partitioning.test > 0, always use Controller,
instead of the prior behavior of only using it when
debug.nn.partitioning.test > 1 and the plan is "in scope"
for execution.

If a combination of the partitioning property value and
the plan's shouldBeExecutable() say the plan shouldn't
be given control, then we do simulated execution under
control of the plan before real execution without using
the plan.  This allows us to test StepExecutors.

("Simulated execution" means iterating through the
StepExecutors without using them to perform any computation.)

Bug: 63905942
Test: mma (user)
      mma (userdebug)
      ml/nn/runtime/tests (userdebug)
      (with debug.nn.partition.test 0, 1, and 2;
       no new failures, logcat looks plausible)

Change-Id: I3a05f5479e906eb13f65e27b17e7129bfcceac0d