Modify test files for size. Add tests.
Modified the specs so that the generated data files won't be
so large as to take very long to compile.

Also generate the tests for the new specs that have been added.

Finally, update the VTS tests for the specs we have.

Bug: 67394766
Test: Compiled and ran unit tests and VTS tests.

The regular tets pass except for:
[  FAILED  ] GeneratedTests.conv_quant8_channels
[  FAILED  ] GeneratedTests.conv_quant8_large
[  FAILED  ] GeneratedTests.conv_quant8_overflow

The VTS tests pass except for:
[  FAILED  ] NeuralnetworksHidlTest.conv_quant8_channels
[  FAILED  ] NeuralnetworksHidlTest.conv_quant8_large
[  FAILED  ] NeuralnetworksHidlTest.conv_quant8_overflow
[  FAILED  ] NeuralnetworksHidlTest.lsh_projection
[  FAILED  ] NeuralnetworksHidlTest.lstm2
[  FAILED  ] NeuralnetworksHidlTest.lstm3
[  FAILED  ] NeuralnetworksHidlTest.lstm

Change-Id: I5a26d0d05e1eca9874b0e9b09f33f345b657667a