Made embedding lookup take int lookups
Bug: 67459279

NN API only supported float tensors when this op was first added.
Now int tensors are allowed in the API and we have updated the
documentation saying the lookup indices are integers, we need
update the implementaion and tests of this op to match the
documentation.

Also reverted the input order to match TF Lite.

Test: NeuralNetworksTest
Test: adb shell
/data/nativetest64/embedding_lookup_test/embedding_lookup_test

Change-Id: Ie0db884ca772e8fb388876c3cef609016e753bf1