Update generated VTS test for mobilenet_quantized
The specification file for mobilenet_quantized was changed
to use a real mobilenet quantized model with a real image
as input. However, the vts test was not regenerated from
that source file when the corresponding change to the spec
file was made. This CL regenerates the vts test for
mobilenet_quantized to keep the two in sync.

Test: mm
Test: VtsHalNeuralnetworksV1_0TargetTest

Merged-In: I2dc7369755dedfafa0fcc1c34611135f6026bd48
Change-Id: I2dc7369755dedfafa0fcc1c34611135f6026bd48
(cherry picked from commit e78b89d46b7e534b749ace863886571a840e83e2)