DO NOT MERGE: Modified the allowable range of precision
modified the allowable range of precision in /frameworks/ml/nn/tools/test_generator/include/TestHarness.h
We change the precision to 0.4% of cts/vts expectations, But with little expectation, a fixed bias is needed
We use 4% of golden expectations as a condition of judgment. But when the golden expectations approaching 0, a fixed bias 0.002 is needed
This CL cannot sufficient  to make all the failed tests ,there still some testcase cannot pass .
Because we calculation using fp16 precision to replace fp32 precision .and some inputs are over fp16  expressive range -32768~+32767
After changing the precision, the mobilenet_quantized  error shows.
We have comfirmed mobilenet_quantized failed because of a bug.
And the others are caused by some inputs  over fp16  expressive range -32768~+32767

Bug:73218205
Test: Build CTS/VTS
Test: EMUI8.1/Android8.1  passed the following tests
Test: run cts -o -m CtsNNAPITestCases
Change-Id: I3158bb121957cbc3c0bcff5a8bd358144512a336