Fix non-threadsafe static gemmlowp::GemmContext
gemmlowp::GemmContext is not threadsafe but was allocated as a
function-level static, leading to memory corruption when multiple CPU
executions were run in parallel.

Fixed by changing to thread_local. This is a minimal change (compared to
creating and passing a state object to cpu implementations) so that the
change is more likely to be accepted for P.

Note that there are some test failures (wrong outputs) in the new tests
that are to be investigated:
[  FAILED  ] GeneratedTests.conv_1_h3_w2_SAME
[  FAILED  ] GeneratedTests.conv_1_h3_w2_SAME_relaxed
[  FAILED  ] GeneratedTests.conv_1_h3_w2_VALID
[  FAILED  ] GeneratedTests.conv_3_h3_w2_SAME
(The new tests are not run as part of any continuous build yet).

Bug: 80430825
Bug: 79931091
Test: NeuralNetworksTest_mt_static/*mobilenet*
Test: NeuralNetworksTest_mt_static_asan/*mobilenet*
Change-Id: Ief11843d5ecfaebb66c7983a8f61e604f55e0469
Merged-In: Ief11843d5ecfaebb66c7983a8f61e604f55e0469
(cherry picked from commit 8c1992b6a1dd716e8e1825cd7c864d3b47bf40c0)